fix: change how unknown routes are handled (fixes a bug on 404) + refactor types + improve/normalise tutorials
