+++ docs/tutorial/jdbc.md
@@ -1,12 +1,16 @@
 ---
-title: Postgres wire client in Java using JDBC
+title: JDBC tutorial
 sidebar_label: JDBC
-description: Tutorial on how to connect to QuestDB using Postgres JDBC driver
+description: Tutorial showing how build a Java application with QuestDB
 ---
 
-Example using the Postgres JDBC driver showing how to connect to QuestDB and execute a prepared statement.
+This tutorial uses the official
+[Postgres JDBC driver](https://jdbc.postgresql.org/) to connect to QuestDB.
 
-Maven dependency:
+## Dependency
+
+### Maven
+
 ```xml
 <dependency>
     <groupId>org.postgresql</groupId>
@@ -15,7 +19,14 @@ Maven dependency:
 </dependency>
 ```
 
-Java class
+### Gradle
+
+```
+compile group: 'org.postgresql', name: 'postgresql', version: '42.2.12'
+```
+
+## Code snippet
+
 ```java
  import java.sql.Connection;
  import java.sql.DriverManager;
@@ -23,9 +34,9 @@ Java class
  import java.sql.ResultSet;
  import java.util.Properties;
  import java.util.TimeZone;
- 
+
  public class JdbcExmple {
-     
+
      public static void main(String[] args) throws Exception {
          Class.forName("org.postgresql.Driver");
          Properties properties = new Properties();
@@ -35,10 +46,10 @@ Java class
          properties.setProperty("binaryTransfer", "true");
          TimeZone.setDefault(TimeZone.getTimeZone("EDT"));
          final Connection connection = DriverManager.getConnection("jdbc:postgresql://127.0.0.1:8812/qdb", properties);
- 
+
          PreparedStatement statement = connection.prepareStatement("select x,? from long_sequence(5)");
          statement.setInt(1, 4);
- 
+
          ResultSet rs = statement.executeQuery();
          rs.close();
          connection.close();
