+++ package.json
@@ -13,26 +13,30 @@
     "serve": "docusaurus serve"
   },
   "dependencies": {
-    "@docusaurus/core": "2.0.0-alpha.61",
-    "@docusaurus/module-type-aliases": "2.0.0-alpha.61",
-    "@docusaurus/plugin-ideal-image": "2.0.0-alpha.61",
-    "@docusaurus/plugin-pwa": "2.0.0-alpha.61",
-    "@docusaurus/preset-classic": "2.0.0-alpha.61",
+    "@docusaurus/core": "2.0.0-alpha.62",
+    "@docusaurus/module-type-aliases": "2.0.0-alpha.62",
+    "@docusaurus/plugin-ideal-image": "2.0.0-alpha.62",
+    "@docusaurus/plugin-pwa": "2.0.0-alpha.62",
+    "@docusaurus/preset-classic": "2.0.0-alpha.62",
     "@questdb/sql-grammar": "1.0.5",
-    "@types/react": "16.9.44",
+    "@tsconfig/docusaurus": "1.0.2",
+    "@types/react": "16.9.49",
+    "@types/react-helmet": "6.1.0",
+    "@types/react-router-config": "5.0.1",
+    "@types/react-router-dom": "5.1.5",
     "@types/throttle-debounce": "2.1.0",
     "clsx": "1.1.1",
-    "date-fns": "2.15.0",
+    "date-fns": "2.16.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "throttle-debounce": "2.2.1",
-    "typescript": "3.9.7",
+    "throttle-debounce": "2.3.0",
+    "typescript": "4.0.2",
     "unist-util-visit": "2.0.3"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "3.8.0",
-    "@typescript-eslint/parser": "3.8.0",
-    "eslint": "7.6.0",
+    "@typescript-eslint/eslint-plugin": "4.0.1",
+    "@typescript-eslint/parser": "4.0.1",
+    "eslint": "7.8.0",
     "eslint-config-prettier": "6.11.0",
     "eslint-config-standard": "14.1.1",
     "eslint-plugin-babel": "5.3.1",
@@ -40,22 +44,22 @@
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-react": "7.20.5",
-    "eslint-plugin-react-hooks": "4.0.8",
+    "eslint-plugin-react": "7.20.6",
+    "eslint-plugin-react-hooks": "4.1.0",
     "eslint-plugin-standard": "4.0.1",
-    "fork-ts-checker-webpack-plugin": "5.0.14",
+    "fork-ts-checker-webpack-plugin": "5.1.0",
     "glob": "7.1.6",
     "graceful-fs": "4.2.4",
     "husky": ">=4",
     "imagemin": "7.0.1",
     "imagemin-gifsicle": "7.0.0",
     "imagemin-svgo": "8.0.0",
-    "lint-staged": ">=10",
+    "lint-staged": "10.2.13",
     "make-dir": "3.1.0",
     "node-fetch": "2.6.0",
-    "prettier": "2.0.5",
+    "prettier": "2.1.1",
     "rimraf": "3.0.2",
-    "sharp": "0.25.4",
+    "sharp": "0.26.0",
     "webpack-manifest-plugin": "2.2.0"
   },
   "browserslist": {
@@ -77,5 +81,8 @@
   },
   "lint-staged": {
     "static/img/**/*": "scripts/check-size"
+  },
+  "resolutions": {
+    "sharp": "0.26.0"
   }
 }
