+++ src/css/instruction.module.css
@@ -2,6 +2,7 @@
   --ifm-link-color: var(--palette-pink);
   --ifm-link-hover-color: var(--palette-pink);
   display: flex;
+  width: 100%;
   max-width: var(--ifm-container-width);
   margin: 0 auto;
   margin-top: 8rem;
