+++ src/theme/DocPage/styles.module.css
@@ -69,6 +69,11 @@ ul ul :global(.table-of-contents__link--active::before) {
   margin-right: initial;
 }
 
+/* @TODO: Hack - to remove when docusaurus .62 is out */
+.doc :global(article > div:first-of-type) {
+  display: none;
+}
+
 .doc__sidebar {
   position: relative;
   width: 300px;
