+++ src/theme/Image/index.tsx
@@ -10,7 +10,6 @@ const IdealImage = ({
     return <img {...props} src={src} />
   }
 
-  // eslint-disable-next-line
   return <Image {...props} />
 }
