+++ tsconfig.json
@@ -33,21 +33,11 @@
     "forceConsistentCasingInFileNames": true,
     "baseUrl": "./",
     "paths": {
-      "@theme/Button": [
-        "src/theme/Button"
-      ],
-      "@theme/Head": [
-        "src/theme/Head"
-      ],
-      "@theme/Image": [
-        "src/theme/Image"
-      ],
-      "@theme/useMetadataContext": [
-        "src/theme/useMetadataContext"
-      ],
-      "@theme/useWindowWidth": [
-        "src/theme/useWindowWidth"
-      ]
+      "@theme/Button": ["src/theme/Button"],
+      "@theme/Head": ["src/theme/Head"],
+      "@theme/Image": ["src/theme/Image"],
+      "@theme/useMetadataContext": ["src/theme/useMetadataContext"],
+      "@theme/useWindowWidth": ["src/theme/useWindowWidth"]
     }
   },
   "include": ["src/", "types/"]
